Created style file, removed original tabbar, created new tab bar
